Add Krita Gemini and Krita Sketch to git master
REVIEW:113327

Conflicts:
	libs/main/KoMainWindow.cpp
	libs/main/KoMainWindow.h

+++ krita/image/tests/kis_fixed_point_maths_test.cpp
@@ -27,8 +27,8 @@ void KisFixedPointMathsTest::testOperators()
 {
     KisFixedPoint fp1(1);
     KisFixedPoint fp2(2);
-    KisFixedPoint fp3(2.5);
-    KisFixedPoint fp4(2.0);
+    KisFixedPoint fp3(qreal(2.5));
+    KisFixedPoint fp4(qreal(2.0));
 
     QVERIFY(fp1 < fp2);
     QVERIFY(fp2 < fp3);
@@ -43,57 +43,57 @@ void KisFixedPointMathsTest::testOperators()
     QCOMPARE(fp1 - fp2, KisFixedPoint(-1));
     QCOMPARE(fp1 * fp2, KisFixedPoint(2));
     QCOMPARE(fp2 * fp3, KisFixedPoint(5));
-    QCOMPARE(fp1 / fp2, KisFixedPoint(0.5));
-    QCOMPARE(fp1 / fp3, KisFixedPoint(0.4));
-    QCOMPARE(fp2 / fp3, KisFixedPoint(0.8));
+    QCOMPARE(fp1 / fp2, KisFixedPoint(qreal(0.5)));
+    QCOMPARE(fp1 / fp3, KisFixedPoint(qreal(0.4)));
+    QCOMPARE(fp2 / fp3, KisFixedPoint(qreal(0.8)));
 }
 
 void KisFixedPointMathsTest::testOperatorsNegative()
 {
-    KisFixedPoint fp1(-2.5);
+    KisFixedPoint fp1(qreal(-2.5));
     KisFixedPoint fp2(2);
     KisFixedPoint fp3(-3);
 
-    QCOMPARE(fp1 + fp2, KisFixedPoint(-0.5));
-    QCOMPARE(fp1 - fp2, KisFixedPoint(-4.5));
+    QCOMPARE(fp1 + fp2, KisFixedPoint(qreal(-0.5)));
+    QCOMPARE(fp1 - fp2, KisFixedPoint(qreal(-4.5)));
     QCOMPARE(fp1 * fp2, KisFixedPoint(-5));
-    QCOMPARE(fp1 * fp3, KisFixedPoint(7.5));
+    QCOMPARE(fp1 * fp3, KisFixedPoint(qreal(7.5)));
 
-    QCOMPARE(fp2 / fp1, KisFixedPoint(-0.8));
-    QCOMPARE(fp1 / fp2, KisFixedPoint(-1.25));
-    QCOMPARE(fp3 / fp1, KisFixedPoint(1.2));
+    QCOMPARE(fp2 / fp1, KisFixedPoint(qreal(-0.8)));
+    QCOMPARE(fp1 / fp2, KisFixedPoint(qreal(-1.25)));
+    QCOMPARE(fp3 / fp1, KisFixedPoint(qreal(1.2)));
 }
 
 void KisFixedPointMathsTest::testConversions()
 {
-    KisFixedPoint fp1(2.5);
-    KisFixedPoint fp2(2.73);
-    KisFixedPoint fp3(2.34);
+    KisFixedPoint fp1(qreal(2.5));
+    KisFixedPoint fp2(qreal(2.73));
+    KisFixedPoint fp3(qreal(2.34));
 
     QCOMPARE(fp1.toInt(), 2);
     QCOMPARE(fp1.toIntRound(), 3);
     QCOMPARE(fp1.toIntFloor(), 2);
     QCOMPARE(fp1.toIntCeil(), 3);
-    QCOMPARE(fp1.toFloat(), 2.5);
+    QCOMPARE(fp1.toFloat(), qreal(2.5));
 
     QCOMPARE(fp2.toInt(), 2);
     QCOMPARE(fp2.toIntRound(), 3);
     QCOMPARE(fp2.toIntFloor(), 2);
     QCOMPARE(fp2.toIntCeil(), 3);
-    QCOMPARE(fp2.toFloat(), 698.0/256.0);
+    QCOMPARE(fp2.toFloat(), qreal(698.0/256.0));
 
     QCOMPARE(fp3.toInt(), 2);
     QCOMPARE(fp3.toIntRound(), 2);
     QCOMPARE(fp3.toIntFloor(), 2);
     QCOMPARE(fp3.toIntCeil(), 3);
-    QCOMPARE(fp3.toFloat(), 599.0/256.0);
+    QCOMPARE(fp3.toFloat(), qreal(599.0/256.0));
 }
 
 void KisFixedPointMathsTest::testConversionsNegative()
 {
-    KisFixedPoint fp1(-2.5);
-    KisFixedPoint fp2(-2.73);
-    KisFixedPoint fp3(-2.34);
+    KisFixedPoint fp1(qreal(-2.5));
+    KisFixedPoint fp2(qreal(-2.73));
+    KisFixedPoint fp3(qreal(-2.34));
 
     QCOMPARE(fp1.toInt(), -2);
     QCOMPARE(fp1.toIntRound(), -3);
